Time and Frames
Just a bit of some old fixing up with some code. I found that since the damage function work on frames, then the value it does to deal damage has to be in a pretty gentle range or else it'll end up being way too hard! Time.deltaTime didn't really help, so I wonder what I can do? It's not that bad anyway.